+++ src/gemseo_umdo/formulations/statistics/taylor_polynomial/standard_deviation.py
@@ -12,16 +12,22 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-"""Formulate a multidisciplinary design problem under uncertainty."""
+"""Estimator of the standard deviation for U-MDO formulations based on Taylor."""
 from __future__ import annotations
 
-from gemseo.formulations.formulations_factory import MDOFormulationsFactory
+from numpy import ndarray
 
-from gemseo_umdo.formulations.formulation import UMDOFormulation
+from gemseo_umdo.formulations.statistics.taylor_polynomial.variance import Variance
 
 
-class UMDOFormulationsFactory(MDOFormulationsFactory):
-    """The factory of U-MDO formulations."""
+class StandardDeviation(Variance):
+    """Estimator of the standard deviation."""
 
-    _CLASS = UMDOFormulation
-    _MODULE_NAMES = ("gemseo_umdo.formulations",)
+    def __call__(self, func: ndarray, jac: ndarray, hess: ndarray) -> ndarray:
+        """
+        Args:
+            func: The output value at the mean value of the uncertain variables.
+            jac: The Jacobian value at the mean value of the uncertain variables.
+            hess: The Hessian value at the mean value of the uncertain variables.
+        """  # noqa: D205 D212 D415
+        return super().__call__(func, jac, hess) ** 0.5