+++ src/gemseo_umdo/formulations/taylor_polynomial.py
@@ -31,11 +31,9 @@ with mean $\mu$ and variance $\sigma^2$.
 from __future__ import annotations
 
 from typing import Any
-from typing import ClassVar
 from typing import Mapping
 from typing import Sequence
 
-from gemseo.algos.database import Database
 from gemseo.algos.design_space import DesignSpace
 from gemseo.algos.doe.lib_custom import CustomDOE
 from gemseo.algos.opt_problem import OptimizationProblem
@@ -43,24 +41,34 @@ from gemseo.algos.parameter_space import ParameterSpace
 from gemseo.core.discipline import MDODiscipline
 from gemseo.core.formulation import MDOFormulation
 from gemseo.core.mdofunctions.mdo_function import MDOFunction
-from gemseo.utils.derivatives.finite_differences import FirstOrderFD
 from gemseo.utils.logging_tools import LoggingContext
-from numpy import atleast_1d
-from numpy import atleast_2d
-from numpy import ndarray
 
-from gemseo_umdo.estimators.taylor_polynomial import (
+from gemseo_umdo.formulations.formulation import UMDOFormulation
+from gemseo_umdo.formulations.functions.hessian_function import (
+    HessianFunction,
+)
+from gemseo_umdo.formulations.functions.statistic_function_for_taylor_polynomial import (
+    StatisticFunctionForTaylorPolynomial,
+)
+from gemseo_umdo.formulations.statistics.taylor_polynomial.taylor_polynomial_estimator_factory import (  # noqa: B950
     TaylorPolynomialEstimatorFactory,
 )
-from gemseo_umdo.formulations.formulation import UMDOFormulation
 
 
 class TaylorPolynomial(UMDOFormulation):
     """Robust MDO formulation based on Taylor polynomials."""
 
-    _STATISTIC_FACTORY: ClassVar[
-        TaylorPolynomialEstimatorFactory
-    ] = TaylorPolynomialEstimatorFactory()
+    __custom_doe: CustomDOE
+    """The DOE to evaluate the MDO formulation at the mean value of the input vector."""
+
+    __hessian_fd_problem: OptimizationProblem | None
+    """The problem related to the approximation of the Hessian if any."""
+
+    __second_order: bool
+    """Whether the formulation uses second-order Taylor polynomials.
+
+    Otherwise first-order Taylor polynomials.
+    """
 
     def __init__(  # noqa: D107
         self,
@@ -76,8 +84,16 @@ class TaylorPolynomial(UMDOFormulation):
         differentiation_method: OptimizationProblem.DifferentiationMethod = OptimizationProblem.DifferentiationMethod.USER_GRAD,  # noqa: B950
         second_order: bool = False,
         **options: Any,
-    ) -> None:
+    ) -> None:  # noqa: D205 D212 D415
+        """
+        Args:
+            differentiation_method: The type of method to compute the gradients.
+            second_order: Whether to use second-order Taylor polynomials
+                instead of first-order Taylor polynomials.
+        """  # noqa: D205 D212 D415
         self.__second_order = second_order
+        self._statistic_function_class = StatisticFunctionForTaylorPolynomial
+        self._statistic_factory = TaylorPolynomialEstimatorFactory()
         super().__init__(
             disciplines,
             objective_name,
@@ -106,7 +122,7 @@ class TaylorPolynomial(UMDOFormulation):
         self.__custom_doe = CustomDOE()
 
     @property
-    def hessian_fd_problem(self) -> OptimizationProblem:
+    def hessian_fd_problem(self) -> OptimizationProblem | None:
         """The problem related to the approximation of the Hessian."""
         return self.__hessian_fd_problem
 
@@ -174,80 +190,3 @@ class TaylorPolynomial(UMDOFormulation):
                 eval_jac=eval_jac,
                 eval_obs_jac=eval_jac,
             )
-
-    class _StatisticFunction(UMDOFormulation._StatisticFunction):
-        def _func(self, input_data: ndarray) -> ndarray:
-            formulation = self._formulation
-            problem = formulation.mdo_formulation.opt_problem
-            if self._function_name in formulation._processed_functions:
-                formulation._processed_functions = []
-                problem.reset()
-                if formulation.hessian_fd_problem is not None:
-                    formulation.hessian_fd_problem.reset()
-
-            database = problem.database
-            if not database:
-                formulation.update_top_level_disciplines(input_data)
-                formulation.evaluate_with_mean(problem, True)
-                if formulation.hessian_fd_problem is not None:
-                    formulation.evaluate_with_mean(
-                        formulation.hessian_fd_problem, False
-                    )
-
-            func_value = atleast_1d(
-                database.get_function_history(self._function_name)[0]
-            )
-            jac_value = atleast_2d(
-                database.get_gradient_history(self._function_name)[0]
-            )
-            hess_value = None
-            if formulation.second_order:
-                hessian_database = formulation.hessian_fd_problem.database
-                hess_name = (
-                    f"{database.GRAD_TAG}{database.GRAD_TAG}{self._function_name}"
-                )
-                hess_value = hessian_database.get_function_history(hess_name)[0]
-                if hess_value.ndim == 1:
-                    hess_value = hess_value[None, None, ...]
-
-                if hess_value.ndim == 2:
-                    hess_value = hess_value[None, ...]
-
-            formulation._processed_functions.append(self._function_name)
-            return self._estimate_statistic(
-                func_value,
-                jac_value,
-                hess_value,
-                **self._statistic_parameters,
-            )
-
-
-class HessianFunction(MDOFunction):
-    """Approximation of the Hessian function with finite differences.
-
-    Take an original function and approximate its Hessian with finite differences
-    applied to its analytical or approximated Jacobian.
-    """
-
-    def __init__(self, func: MDOFunction) -> None:
-        """
-        Args:
-            func: The original function.
-        """  # noqa: D205 D212 D415
-        self.__jac = func.jac if func.has_jac else FirstOrderFD(func.func).f_gradient
-        grad_tag = Database.GRAD_TAG
-        super().__init__(
-            FirstOrderFD(self._compute_jac).f_gradient,
-            f"{grad_tag}{grad_tag}{func.name}",
-        )
-
-    def _compute_jac(self, input_data: ndarray) -> ndarray:
-        """Compute the Jacobian matrix.
-
-        Args:
-            input_data: The input data.
-
-        Returns:
-            The Jacobian matrix.
-        """
-        return self.__jac(input_data).T