+++ tests/formulations/test_taylor_polynomial.py
@@ -21,10 +21,12 @@ from gemseo.algos.design_space import DesignSpace
 from gemseo.algos.parameter_space import ParameterSpace
 from gemseo.core.discipline import MDODiscipline
 from gemseo.formulations.mdf import MDF
-from gemseo_umdo.estimators.taylor_polynomial import Margin
-from gemseo_umdo.estimators.taylor_polynomial import Mean
-from gemseo_umdo.estimators.taylor_polynomial import StandardDeviation
-from gemseo_umdo.estimators.taylor_polynomial import Variance
+from gemseo_umdo.formulations.statistics.taylor_polynomial.margin import Margin
+from gemseo_umdo.formulations.statistics.taylor_polynomial.mean import Mean
+from gemseo_umdo.formulations.statistics.taylor_polynomial.standard_deviation import (
+    StandardDeviation,
+)
+from gemseo_umdo.formulations.statistics.taylor_polynomial.variance import Variance
 from gemseo_umdo.formulations.taylor_polynomial import TaylorPolynomial
 from gemseo_umdo.scenarios.udoe_scenario import UDOEScenario
 from numpy import array
@@ -77,9 +79,7 @@ def umdo_formulation_with_hessian(
 
 
 @pytest.mark.parametrize("second_order", [False, True])
-def test_scenario(
-    disciplines, design_space, uncertain_space, second_order, mdf_discipline, tmp_path
-):
+def test_scenario(disciplines, design_space, uncertain_space, second_order, tmp_path):
     """Check the optimum returned by the UMDOScenario."""
     scn = UDOEScenario(
         disciplines,
@@ -95,13 +95,15 @@ def test_scenario(
     scn.add_observable("o", "Variance")
     file_path = tmp_path / "scenario.h5"
     scn.to_pickle(file_path)
-    algo_data = {"algo": "CustomDOE", "algo_options": {"samples": array([[0.0] * 3])}}
+    algo_data = {
+        "algo": "CustomDOE",
+        "algo_options": {"samples": array([[0.0, 0.0, 0.0], [1.0, 1.0, 1.0]])},
+    }
     scn.execute(algo_data)
     saved_scn = UDOEScenario.from_pickle(file_path)
     saved_scn.execute(algo_data)
-    uncertain_data = uncertain_space.array_to_dict(uncertain_space.distribution.mean)
-    expected_output_data = mdf_discipline.execute(uncertain_data)["f"]
-    expected_input_data = array([0.0] * 3)
+    expected_output_data = array([-21.0])
+    expected_input_data = array([1.0] * 3)
     assert_equal(scn.optimization_result.x_opt, expected_input_data)
     assert_equal(scn.optimization_result.f_opt, expected_output_data)
     assert_equal(saved_scn.optimization_result.x_opt, expected_input_data)
@@ -128,25 +130,29 @@ def test_init_differentiation_method(umdo_formulation):
 
 def test_estimate_mean(umdo_formulation):
     """Check that E[f(x,U)] is estimated by f(x,m) + 0.5s²h(x,m)."""
-    mean_estimation = Mean(umdo_formulation)(FUNC, JAC, HESS)
+    mean_estimation = Mean(umdo_formulation.uncertain_space)(FUNC, JAC, HESS)
     assert_equal(mean_estimation, array([115.0, -112.0]))
 
 
 def test_estimate_variance(umdo_formulation):
     """Check that V[f(x,U)] is estimated by (sj(x,m))²."""
-    var_estimation = Variance(umdo_formulation)(FUNC, JAC, HESS)
+    var_estimation = Variance(umdo_formulation.uncertain_space)(FUNC, JAC, HESS)
     assert_equal(var_estimation, array([98.0, 440.0]))
 
 
 def test_estimate_standard_derivation(umdo_formulation):
     """Check that V[f(x,u)] = S[f(x,U)]²."""
-    std_estimation = StandardDeviation(umdo_formulation)(FUNC, JAC, HESS)
+    std_estimation = StandardDeviation(umdo_formulation.uncertain_space)(
+        FUNC, JAC, HESS
+    )
     assert_equal(std_estimation, array([98.0, 440.0]) ** 0.5)
 
 
 def test_estimate_margin(umdo_formulation):
     """Check that Margin[f(x,u);k] = E[f(x,U)] + kS[f(x,U)]."""
-    margin_estimation = Margin(umdo_formulation)(FUNC, JAC, hess=HESS, factor=3.0)
+    margin_estimation = Margin(umdo_formulation.uncertain_space, factor=3.0)(
+        FUNC, JAC, HESS
+    )
     assert_equal(
         margin_estimation, array([115.0, -112.0]) + 3.0 * array([98.0, 440.0]) ** 0.5
     )